Refactorización del código y modularización
