+++ module.path.js
@@ -1,6 +1,6 @@
-const path = require("path");
+// const path = require("path");
 
-let pathObj = path.resolve("README.md");
-//let pathObj = path.extname(__filename);
-console.log(pathObj);
-// console.log(pathObj.ext);
+// let pathObj = path.resolve("README.md");
+// //let pathObj = path.extname(__filename);
+// console.log(pathObj);
+// // console.log(pathObj.ext);
